// src/app/models/hotel.ts
export class Hotel {
    operation?: string;
    nombre?: string;
    nit?: string;
    rnt?: string;
    matriculaMercantil?: string;
    direccion?: string;
    email?: string;
    telefono?: string;
    nombreRepLegal?: string;
    telefonoRepLegal?: string;
    emailRepLegal?: string;
    nombreContacto?: string;
    telefonoContacto?: string;
    idMunicipio?: number;
}